LayoutTests/http/tests/security/xssAuditor/full-block-object-tag-expected.txt

CONSOLE MESSAGE: line 1: Refused to load an object. URL found within request: "http://localhost:8000/security/xssAuditor/resources/dummy.swf".

There should be no content in the iframe below:



--------
Frame: 'frame'
--------
